requirements.txt
django>=5.1.2,<5.2
psycopg2>=2.9.9,<3.0
djangorestframework>=3.15.2,<3.16
djangorestframework-gis>=1.1,<1.2
django-cors-headers>=4.5.0,<4.6